tests/fixtures/django-app/requirements.txt
django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.15.0
celery==5.4.0
redis==5.0.1
django-allauth==0.61.0
django-environ==0.11.2
pytest==8.0.0
pytest-django==4.8.0
ruff==0.2.0
mypy==1.8.0
factory-boy==3.3.0
coverage==7.4.0
gunicorn==21.2.0
